Fix MSVC warnings
Explicit casts are added around expressions assigned to variables of
narrower integral types (ptrdiff_t -> unsigned int, int -> char).

A constant char array (the usage synopsis line) is redeclared as a
macro to silence the warning complaining that the format arg to fprintf
is not a string literal.

+++ src/grid_io.c
@@ -147,7 +147,7 @@ static inline int _init_grid_from_plain(struct grid *grid, const char *repr,
 	   just check that all line lengths are equal *including* the CR, and
 	   decrement the width to account for CRs only once at the end. */
 	bool is_crlf = *(first_nl - 1) == '\r';
-	unsigned int width = first_nl - repr;
+	unsigned int width = (unsigned int) (first_nl - repr);
 	unsigned int height = 1;
 	const char *next_nl;
 	while ((next_nl = strchr(repr, '\n')) != NULL) {
